ensure integer on numer format

+++ branches/2.x/application/lib/framework/Parser.php
@@ -517,7 +517,7 @@ class Xerxes_Framework_Parser
 	
 	public static function number_format($number, $decimals = 0)
 	{
-		$number = preg_replace('/\D/', '', $number);
+		$number = (int) preg_replace('/\D/', '', $number);
 		
 		$localeconv = localeconv();
